requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.14.0
altair==5.4.1
anyio==4.6.2.post1
apache-airflow==2.10.3
apache-airflow-providers-common-compat==1.2.1
apache-airflow-providers-common-io==1.4.2
apache-airflow-providers-common-sql==1.19.0
apache-airflow-providers-fab==1.5.0
apache-airflow-providers-ftp==3.11.1
apache-airflow-providers-http==4.13.2
apache-airflow-providers-imap==3.7.0
apache-airflow-providers-smtp==1.8.0
apache-airflow-providers-sqlite==3.9.0
apispec==6.7.1
argcomplete==3.5.1
asgiref==3.8.1
attrs==24.2.0
babel==2.16.0
blinker==1.8.2
cachelib==0.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==6.9.0
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.5
croniter==5.0.1
cryptography==43.0.3
Deprecated==1.2.14
dill==0.3.9
dnspython==2.6.1
docopt==0.6.2
email_validator==2.2.0
Flask==2.2.5
Flask-AppBuilder==4.5.2
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
google-re2==1.1.20240702
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.1
gunicorn==23.0.0
h11==0.14.0
hdfs==2.7.3
httpcore==1.0.6
httpx==0.27.2
idna==3.8
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kafka-python==2.0.2
kafka-python-ng==2.2.2
lazy-object-proxy==1.10.0
limits==3.13.0
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.5.0
multidict==6.1.0
narwhals==1.8.1
numpy==2.1.1
opentelemetry-api==1.28.0
opentelemetry-exporter-otlp==1.28.0
opentelemetry-exporter-otlp-proto-common==1.28.0
opentelemetry-exporter-otlp-proto-grpc==1.28.0
opentelemetry-exporter-otlp-proto-http==1.28.0
opentelemetry-proto==1.28.0
opentelemetry-sdk==1.28.0
opentelemetry-semantic-conventions==0.49b0
ordered-set==4.1.0
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
pendulum==3.0.0
pillow==10.4.0
pluggy==1.5.0
prison==0.2.1
propcache==0.2.0
protobuf==5.28.1
psutil==6.1.0
psycopg2-binary==2.9.10
pyarrow==17.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
pymongo==4.8.0
pytest==8.3.3
pytest-mock==3.14.0
python-daemon==3.1.0
python-dateutil==2.9.0.post0
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.8.1
rich-argparse==1.6.0
rpds-py==0.20.0
setproctitle==1.3.3
setuptools==75.3.0
six==1.16.0
smmap==5.0.1
snakebite==2.11.0
sniffio==1.3.1
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.1
streamlit==1.38.0
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.5.0
text-unidecode==1.3
time-machine==2.16.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
universal_pathlib==0.2.5
urllib3==2.2.2
watchdog==4.0.2
Werkzeug==2.2.3
wirerope==0.4.7
wrapt==1.16.0
WTForms==3.2.1
yarl==1.17.1
zipp==3.20.2
